소수점 입력 부분 재정의, 코드 정리
코드정리
1. main.dart
2. editRecipe.dart
3. addRecipe.dart

+++ lib/controller/recipe_controller.dart
@@ -49,10 +49,10 @@ class RecipeController extends GetxController{
   }
   multipleValueUpdate(menuTitle, recipeTitle, index, multipleIndicator)async{
     await firestore.collection('users').doc(email).collection(menuTitle).doc('Recipe').update(
-      // {recipeTitle:{'multipleValue':int.parse(multipleIndicator), 'ingredient':recipeIngredient[index], 'weight':recipeWeight[index]}}
-      {recipeTitle:{'multipleValue':double.parse(multipleIndicator), 'ingredient':recipeIngredient[index], 'weight':recipeWeight[index]}}
+      {recipeTitle:{'multipleValue':int.parse(multipleIndicator), 'ingredient':recipeIngredient[index], 'weight':recipeWeight[index]}}
+      // {recipeTitle:{'multipleValue':double.parse(multipleIndicator), 'ingredient':recipeIngredient[index], 'weight':recipeWeight[index]}}
     );
-    // multipleValue[index] = int.parse(multipleIndicator);
-    multipleValue[index] = double.parse(multipleIndicator);
+    multipleValue[index] = int.parse(multipleIndicator);
+    // multipleValue[index] = double.parse(multipleIndicator);
   }
 }
